fix(web): rename pill helpers to avoid case-insensitive path clash
macOS resolves ComposerModelPill.tsx and composerModelPill.ts to the same import, which broke the Vite build.
